requirements.txt
twitter
Django==1.6.5
jsonfield==0.9.20
django-suit==0.2.5
south==0.8.4
Pillow==2.4.0
docutils==0.11
django-rosetta==0.7.3
django-pipeline==1.3.23
django-debug-toolbar==1.0.1
django-allauth==0.15.0
djangorestframework==2.3.13
django-extensions==1.3.3
psycopg2==2.5.2
Pygments==1.6
django-filter==0.7
django-autoslug==1.7.2
Markdown==2.4.1
metron==1.3.1
raven
django-braces==1.3.1
django-markdown
django-taggit==0.12.0
Unidecode==0.04.16
